[doc-build] Merge with master

+++ numerics/src/FrictionContact/gfc3d_compute_error.c
@@ -28,6 +28,7 @@
 #include "numerics_verbose.h"              // for numerics_error, numerics_w...
 #include "sanitizer.h"                     // for cblas_dcopy_msan
 #include "SiconosBlas.h"                         // for cblas_dcopy, cblas_dnrm2
+#include "projectionOnCone.h"
 
 int gfc3d_compute_error(GlobalFrictionContactProblem* problem,
                         double*  reaction , double*  velocity,
@@ -177,7 +178,7 @@ int gfc3d_compute_error_convex(GlobalFrictionContactProblem* problem,
   /* Computes error = dnorm2( GlobalVelocity -M^-1( q + H reaction)*/
   int nc = problem->numberOfContacts;
   int m = nc * 3;
-  int n = problem->M->size0;
+  size_t n = problem->M->size0;
   double *mu = problem->mu;
   double *q = problem->q;
